requirements.txt
Flask>=3.0
mysql-connector-python>=9.0
openpyxl>=3.1
